+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'boardspace'
+rootProject.name = 'oraclejpa'
