--- src/main/java/com/boardspace/repository/MemoryPostRepository.java
@@ -1,7 +1,0 @@
-package com.boardspace.repository;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class MemoryPostRepository {
-}
