+++ src/main/java/com/oraclejpa/controller/PostController.java
@@ -1,8 +1,8 @@
-package com.boardspace.controller;
+package com.oraclejpa.controller;
 
-import com.boardspace.model.Post;
-import com.boardspace.model.User;
-import com.boardspace.service.PostService;
+import com.oraclejpa.model.Post;
+import com.oraclejpa.model.User;
+import com.oraclejpa.service.PostService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -67,7 +67,7 @@ public class PostController {
         Post post = postService.findById(id);
         model.addAttribute("post", post);
         model.addAttribute("page", page);
-        return "post/detail-v1";
+        return "post/detail";
     }
 
     @GetMapping("/update/{id}")
