+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ spring.datasource.username=mini
 spring.datasource.password=2417
 spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
 spring.liquibase.change-log=classpath:/db/changelog/db.changelog-master.yaml
-spring.liquibase.enabled=false
+spring.liquibase.enabled=true
 
 #spring.jpa.show-sql=true
 #spring.jpa.properties.hibernate.format_sql=true
