+++ src/main/resources/templates/post/articles-v1.html
@@ -4,37 +4,63 @@
     <meta charset="UTF-8">
     <title>글 목록 페이지</title>
 
-    <link rel="stylesheet" th:href="@{/css/styles-v1.css}" href="../../static/css/styles-v1.css">
+    <link rel="stylesheet" th:href="@{/css/styles.css}" href="../../static/css/styles.css">
+
+    <style>
+        /* Background-image Styles */
+        body {
+            background-image: url(/image/milkyway.jpg);
+            background-size: 1920px 1620px;
+        }
+
+        /* Link Styles */
+        a {
+            color: #bebebe;
+            padding: 15px;
+        }
+
+        a:hover {
+            color: #ffffff;
+        }
+
+        a:focus {
+            color: #ffffff;
+        }
+    </style>
 </head>
-<body class="bg-image-milkyway">
+<body>
     <h1>글 목록</h1>
 
-    <div class="flex-col align-center mt-5vh">
-        <div class="div-width-50p" th:each="post : ${postPage.content}">
+    <div style="display: flex; flex-direction: column; align-items: center;">
+        <div th:each="post : ${postPage.content}" style="width: 800px;">
             <form th:action="@{/post/detail/{id}(id=${post.id})}" method="get">
                 <input type="hidden" th:value="${postPage.number + 1}" name="page" />
-                <div class="post-list" onclick="this.closest('form').submit();">
-                    <div class="p-5 my-10 font-size-16">
-                        <p th:text="${post.nickname}"></p>
+                <div class="post-box" onclick="this.closest('form').submit();">
+                    <div style="cursor: pointer; margin-left: 10px;">
+                        <div style="display: flex; gap: 50px; margin-bottom: 5px;">
+                            <p th:text="${post.nickname}"></p>
+                        </div>
                         <p th:text="${post.title}"></p>
                     </div>
+
+                    <hr style="margin: 15px 0;">
                 </div>
-                <hr>
             </form>
         </div>
 
-        <div class="flex mt-10 font-size-16">
+        <div style="color: aliceblue; font-size: 16px; display: flex;">
             <div th:each="pageNum : ${#numbers.sequence(1, postPage.totalPages)}">
-                <div class="page-line">
-                    <a class="inactive-page" th:if="${pageNum != postPage.number + 1}" th:text="${pageNum}" th:href="@{/post/articles(page=${pageNum})}"></a>
-                    <span class="active-page" th:if="${pageNum == postPage.number + 1}" th:text="${pageNum}"></span>
-                </div>
+                <a th:if="${pageNum != postPage.number + 1}" th:text="${pageNum}" th:href="@{/post/articles(page=${pageNum})}"></a>
+                <span th:if="${pageNum == postPage.number + 1}" class="no-pointer" th:text="${pageNum}" style="color: #659bff; padding: 15px;"></span>
             </div>
         </div>
 
         <div class="button-container">
-            <button class="btn btn-primary" type="button" onclick="window.location.href='/post/write'">글 작성</button>
-            <button class="btn btn-primary" type="button" onclick="window.location.href='/post'">돌아가기</button>
+            <form th:action="@{/post/write}" method="get">
+                <button type="submit">글 작성</button>
+            </form>
+
+            <button type="button" onclick="window.location.href='/post'">돌아가기</button>
         </div>
     </div>
 </body>
