+++ src/main/resources/templates/post/detail.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>글 조회 페이지</title>
+    <script>
+        function confirmDeletePost() {
+            let confirmFlag = confirm("정말 삭제하시겠습니까?");
+            if(confirmFlag) {
+                alert("글이 성공적으로 삭제되었습니다.")
+                return true;
+            } else {
+                return false;
+            }
+        }
+    </script>
+</head>
+<body>
+    <h1 style="color: gray">게시글 조회</h1>
+    <div th:text="'ID: ' + ${post.id}"></div>
+    <div th:text="'제목: ' + ${post.title}"></div>
+    <div th:text="'내용: ' + ${post.content}"></div>
+
+    <div class="button-container" style="display:flex; gap:10px;">
+        <form th:action="@{/post/update/{id}(id=${post.id})}" method="get">
+            <input type="hidden" th:value="${page}" name="page" />
+            <button type="submit">수정</button>
+        </form>
+
+        <form th:action="@{/post/delete}" method="post" onsubmit="return confirmDeletePost()">
+            <input type="hidden" th:value="${post.id}" name="id">
+            <button type="submit">삭제</button>
+        </form>
+
+        <form th:action="@{/post/articles}" method="get">
+            <input type="hidden" th:value="${page}" name="page">
+            <button type="submit">글 목록</button>
+        </form>
+    </div>
+</body>
+</html>
