+++ src/main/resources/templates/user/index-v1.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>홈 화면</title>
 
-    <link rel="stylesheet" th:href="@{/css/styles-v1.css}" href="../../static/css/styles-v1.css">
+    <link rel="stylesheet" th:href="@{/css/styles.css}" href="../../static/css/styles.css">
 
     <style>
         label[for="remember"] {
@@ -21,7 +21,7 @@
         }
     </script>
 </head>
-<body class="bg-image-flower">
+<body>
     <h1>홈페이지</h1>
 
     <div th:if="${logInMessage != null}">
@@ -48,12 +48,12 @@
 
             <tr>
                 <td colspan="2">
-                    <div class="flex-center">
-                        <label class="m-0 align-center" for="remember">아이디 기억</label>
+                    <div class="center">
+                        <label for="remember">아이디 기억</label>
                         <input type="checkbox" id="remember" name="remember"
                             style="width: auto; margin: auto 20px auto 5px;">
 
-                        <button class="btn btn-primary" type="submit">로그인</button>
+                        <button type="submit" style="margin-right: 120px;">로그인</button>
                     </div>
                 </td>
             </tr>
@@ -63,9 +63,11 @@
             <td colspan="2">
                 <div class="button-container">
                     <form th:action="@{/signup}" method="get">
-                        <button class="btn btn-primary" type="submit">회원가입</button>
+                        <button type="submit">회원가입</button>
                     </form>
-                    <button class="btn btn-primary" onclick="window.location.href='/forgot-id'">아이디 / 비밀번호 찾기</button>
+                    <form th:action="@{/find-data}" method="get">
+                        <button type="submit">아이디 / 비밀번호 찾기</button>
+                    </form>
                 </div>
             </td>
         </tr>
