+++ src/main/resources/templates/user/my-page-v1.html
@@ -1,48 +1,56 @@
 <!DOCTYPE html>
 <html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <title>유저 정보 수정 페이지</title>
-        <link rel="stylesheet" th:href="@{/css/styles-v1.css}" href="../../static/css/styles-v1.css">
+<head>
+    <meta charset="UTF-8">
+    <title>유저 정보 수정 페이지</title>
+    <link rel="stylesheet" th:href="@{/css/styles.css}" href="../../static/css/styles.css">
 
-        <script>
-            function showEditUserAlert() {
-                alert("유저 정보가 변경되었습니다.")
-            }
-        </script>
-    </head>
-    <body class="bg-image-milkyway">
-        <h1>회원정보</h1>
-        <div class="flex">
-            <div class="flex-column">
-                <form th:action="@{/post/my-post}" method="get">
-                    <button class="btn btn-primary" type="submit">내 글</button>
-                </form>
-                <button class="btn btn-primary" type="button">내 댓글</button>
-                <form th:action="@{/withdrawal}" method="get">
-                    <button class="btn btn-secondary" type="submit">회원탈퇴</button>
-                </form>
+    <style>
+        label {
+            display: flex;
+            font-size: 14px;
+            margin-left: 3px;
+            margin-bottom: 3px;
+        }
+        input::placeholder {
+            font-size: 14px;
+        }
+    </style>
+
+    <script>
+        function showEditUserAlert() {
+            alert("유저 정보가 변경되었습니다.")
+        }
+    </script>
+</head>
+<body>
+    <h1>회원정보</h1>
+    <div class="center">
+        <form th:action="@{/user/update}" th:object="${user}" method="post" onsubmit="showEditUserAlert()">
+            <label for="userId">아이디</label>
+            <input class="no-pointer" type="text" id="userId" th:value="${user.userId}" th:field="*{userId}" style="background: #818181;" readonly />
+
+            <label for="email">이메일</label>
+            <input type="email" id="email" th:value="${user.email}" th:field="*{email}"
+                placeholder="이메일을 입력해주세요." autofocus />
+        
+            <label for="nickname">닉네임</label>
+            <input type="text" id="nickname" th:value="${user.nickname}" th:field="*{nickname}"
+                placeholder="닉네임을 입력해주세요." />
+        
+            <div class="button-container">
+                <button type="submit">저장</button>
+                <button type="button" onclick="window.location.href='/'">돌아가기</button>
             </div>
+        </form>
 
-            <div class="m-0-auto">
-                <form th:action="@{/user/update}" th:object="${user}" method="post" onsubmit="showEditUserAlert()">
-                    <label for="userId">아이디</label>
-                    <input class="no-pointer input-width-400" type="text" id="userId" th:value="${user.userId}" th:field="*{userId}" style="background: #818181;" readonly />
-
-                    <label for="email">이메일</label>
-                    <input class="input-width-400" type="email" id="email" th:value="${user.email}" th:field="*{email}"
-                        placeholder="이메일을 입력해주세요." autofocus />
-                
-                    <label for="nickname">닉네임</label>
-                    <input class="input-width-400" type="text" id="nickname" th:value="${user.nickname}" th:field="*{nickname}"
-                        placeholder="닉네임을 입력해주세요." />
-                
-                    <div class="flex-center gap-20">
-                        <button class="btn btn-primary" type="submit">저장</button>
-                        <button class="btn btn-primary" type="button" onclick="window.location.href='/'">돌아가기</button>
-                    </div>
-                </form> 
-            </div>
-        </div>
-    </body>
+        <form th:action="@{/post/my-post}" method="get">
+            <button type="submit">내 글</button>
+        </form>
+        <button type="button">내 댓글</button>
+        <form th:action="@{/withdrawal}" method="get">
+            <button type="submit">회원탈퇴</button>
+        </form>
+    </div>
+</body>
 </html>
