+++ src/main/resources/templates/user/my-post-v1.html
@@ -3,10 +3,15 @@
 <head>
     <meta charset="UTF-8">
     <title>내 글 페이지</title>
-
-    <link rel="stylesheet" th:href="@{/css/styles-v1.css}" href="../../static/css/styles-v1.css">
+    <link rel="stylesheet" th:href="@{/css/styles.css}" href="../../static/css/styles.css">
 
     <style>
+        /* Background-image Styles */
+        body {
+            background-image: url(/image/milkyway.jpg);
+            background-size: 1920px 1620px;
+        }
+
         /* Link Styles */
         a {
             color: #bebebe;
@@ -22,14 +27,14 @@
         }
     </style>
 </head>
-<body class="bg-image-milkyway">
+<body>
     <h1>글 목록</h1>
 
-    <div class="flex-col" style="align-items: center;">
+    <div style="display: flex; flex-direction: column; align-items: center;">
         <div th:each="post : ${postPage.content}" style="width: 800px;">
             <form th:action="@{/post/detail/{id}(id=${post.id})}" method="get">
                 <input type="hidden" th:value="${postPage.number + 1}" name="page" />
-                <div id="articles" onclick="this.closest('form').submit();">
+                <div class="post-box" onclick="this.closest('form').submit();">
                     <div style="cursor: pointer; margin-left: 10px;">
                         <div style="display: flex; gap: 50px; margin-bottom: 5px;">
                             <p th:text="${post.nickname}"></p>
@@ -51,10 +56,10 @@
 
         <div class="button-container">
             <form th:action="@{/post/write}" method="get">
-                <button class="btn btn-primary" type="submit">글 작성</button>
+                <button type="submit">글 작성</button>
             </form>
 
-            <button class="btn btn-primary" type="button" onclick="window.location.href='/post'">돌아가기</button>
+            <button type="button" onclick="window.location.href='/post'">돌아가기</button>
         </div>
     </div>
 </body>
