--- src/main/resources/templates/user/reset-pw.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <title>비밀번호 변경 페이지</title>
-
-        <style>
-            body {
-                overflow: hidden;
-            }
-
-            .check-message {
-                font-size: 14px;
-                text-align: start;
-            }
-
-            #password-check {
-                margin-bottom: 5px;
-            }
-        </style>
-
-        <link rel="stylesheet" th:href="@{/css/styles-v1.css}" href="../../static/css/styles-v1.css">
-
-        <script src="http://code.jquery.com/jquery-latest.js"></script>
-        <script src="/js/user.js"></script>
-    </head>
-    <body class="bg-image-flower">
-        <div class="button-container mt-5p">
-            <button class="btn btn-unselected" onclick="window.location.href='/forgot-id'">아이디 찾기</button>
-            <button class="btn btn-selected">비밀번호 찾기</button>
-        </div>
-
-        <div class="div-width-400 m-0-auto">
-            <div th:if="${id}">
-                <p class="mt-20p">새로운 비밀번호를 입력해주세요.</p>
-                <p>기존의 비밀번호와 동일한 비밀번호로는 초기화하실 수 없습니다.</p>
-                <form th:action="reset-pw" method="post">
-                    <div class="flex-col mt-10p">
-                        <div class="m-0-auto div-width-400">
-                            <label for="password">비밀번호</label>
-                            <input class="input-width-100p" type="password" id="password" name="password" maxlength="12">
-                            <label for="password-check">비밀번호 확인</label>
-                            <input class="input-width-100p" type="password" id="password-check" name="password-check" maxlength="12">
-                            <p class="check-message"></p>
-                        </div>
-                    </div>
-                    
-                    <div class="button-container">
-                        <button class="btn btn-primary" type="button" onclick="window.location.href='/'">돌아가기</button>
-                        <button class="btn btn-primary-disabled" type="submit" id="change">비밀번호 변경</button>
-                    </div>
-
-                    <div>
-                        <p class="mt-20">문의하실 사항이 있을 경우 고객센터로 연락해주시기 바랍니다.</p>
-                        <p>고객센터: (02) 123-1234</p>
-                    </div>
-                </form>
-            </div>
-
-            <div th:if="${id == null}">
-                <p class="mt-30p">해당 정보로 가입된 계정이 존재하지 않습니다.</p>
-                <p>아이디 또는 이메일을 다시 확인하시거나 회원가입을 진행해주세요.</p>
-                <p class="mt-20">문의하실 사항이 있을 경우 고객센터로 연락해주시기 바랍니다.</p>
-                <p>고객센터: (02) 123-1234</p>
-
-                <div class="button-container mt-5p">
-                    <button class="btn btn-primary" onclick="window.location.href='/'">로그인</button>
-                    <button class="btn btn-primary" onclick="window.location.href='/signup'">회원가입</button>
-                </div>
-            </div>
-        </div>
-    </body>
-</html>
