--- src/main/resources/templates/user/return-id.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <title>아이디 찾기 페이지</title>
-
-        <style>
-            body {
-                overflow: hidden;
-            }
-        </style>
-
-        <link rel="stylesheet" th:href="@{/css/styles-v1.css}" href="../../static/css/styles-v1.css">
-    </head>
-    <body class="bg-image-flower">
-        <div class="button-container mt-5p">
-            <button class="btn btn-selected">아이디 찾기</button>
-            <button class="btn btn-unselected" onclick="window.location.href='/forgot-pw'">비밀번호 찾기</button>
-        </div>
-
-        <div class="div-width-400 m-0-auto">
-            <div th:if="${userid}">
-                <p class="mt-30p">아이디 찾기가 완료되었습니다.</p>
-                <p>해당 이메일로 가입된 아이디는 다음과 같습니다.</p>
-                <input class="flex mt-10 m-0-auto input-width-200" type="text" th:value="${userid}" disabled>
-                <p class="mt-10p">만약 본인이 가입하신 아이디가 아닐 경우, 고객센터로 연락해주시기 바랍니다.</p>
-                <p>고객센터: (02) 123-1234</p>
-            </div>
-
-            <div th:unless="${userid}">
-                <p class="mt-30p">해당 이메일로 가입된 아이디가 존재하지 않습니다.</p>
-                <p>이메일을 다시 확인하시거나 회원가입을 진행해주세요.</p>
-                <div class="mt-5p">
-                    <p>문의하실 사항이 있을 경우 고객센터로 연락해주시기 바랍니다.</p>
-                    <p>고객센터: (02) 123-1234</p>
-                </div>
-            </div>
-
-            <div class="button-container">
-                <button class="btn btn-primary" onclick="window.location.href='/'">로그인</button>
-                <button class="btn btn-primary" onclick="window.location.href='/signup'">회원가입</button>
-            </div>
-        </div>
-    </body>
-</html>
