+++ src/main/resources/templates/user/signup-v1.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>회원가입</title>
 
-    <link rel="stylesheet" th:href="@{css/styles-v1.css}" href="../../static/css/styles-v1.css">
+    <link rel="stylesheet" th:href="@{css/styles.css}" href="../../static/css/styles.css">
 
     <script>
         function showSignUpAlert() {
@@ -12,7 +12,7 @@
         }
     </script>
 </head>
-<body class="bg-image-flower">
+<body>
     <h1>회원가입 페이지</h1>
     <table>
         <form th:action="@{/signup}" th:object="${user}" method="post" onsubmit="showSignUpAlert()">
@@ -47,8 +47,8 @@
             <tr>
                 <td colspan="2">
                     <span class="button-container">
-                        <button class="btn btn-primary" type="submit">회원가입</button>
-                        <button class="btn btn-primary" type="button" onclick="window.location.href='/'">돌아가기</button>
+                        <button type="submit">회원가입</button>
+                        <button type="button" onclick="window.location.href='/'">돌아가기</button>
                     </span>
                 </td>
             </tr>
